Bully-Algorithm implementation
Bully algorithm is an election algorithm which aim to find the node with maximum UUID and set it as coordinator for all nodes with victory and election messages.
When the simulation starts the first node checks whether it has the biggest UUID or not the it sends "Victory" message if it has the biggest. If it is not the biggest the it sends "Election" message to nodes which have bigger UUID then it. When some node sends victory message simulation stops and every node sets coordinator. If you change N parameter you can decide how many nodes will be in the simulator.